feat: PRism 종합 리포트 컴포넌트 분리 (#80)
